requirements.txt
anthropic==0.51.0
antlr4-python3-runtime==4.7.2
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
google==3.0.0
google-auth==2.39.0
google-genai==1.13.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jinja2==3.1.5
jiter==0.8.2
lexicon==1.0.0
lightning==2.5.0.post0
markupsafe==3.0.2
more-itertools==10.6.0
multipledispatch==1.0.0
numpy==1.26.0
openai==1.63.1
pillow==9.5.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.1
pyperplan==2.1
pyprover==0.6.2
pytamer==0.1.21
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
scikit-image==0.24.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tarski==0.8.2
torch==2.6.0
unified-planning==1.2.0
up-aries==0.4.0
up-enhsp==0.0.27
up-fast-downward==0.5.0
up-fmap==0.0.13
up-lpg==0.1.2
up-pyperplan==1.1.0
up-symk==1.3.1
up-tamer==1.1.5
websockets==15.0.1

# Notes
# - Heavy binary packages (torch, up-symk and other up-* backends) are included here
#   to match the original project environment. If you prefer a lightweight install,
#   remove or comment out torch and the up-* planner packages and install them via conda.
matplotlib==3.10.7
omegaconf==2.3.0
hydra-core==1.3.2
minigrid
gymnasium==0.30.1

# Simulator Requirements


# AlfWorld (optional - install separately if needed)
# alfworld @ git+https://github.com/alfworld/alfworld.git

# VirtualHome (optional - install separately if needed)
# virtualhome @ git+https://github.com/xavierpuigf/virtualhome.git

# Utilities
python-dotmap>=1.3.30
